refactor(web): relocate GitHub-drift subtree to lib/github/ (SP2-C Task 1)
Move github.ts, publish-pin.ts, drift.ts, content-commit.ts + their tests
into lib/github/. Split the shared content-sync/types.ts: ChecksState,
GitHubUnavailableError, BlockedCommitError, MaskMismatchError, RepoTree now
live in lib/github/types.ts. Pure relocation — no behavior change.

Repointed importers: content/{drift,sync}/route.ts, courses/sync/route.ts,
publish/pin/{route,client}, admin-signer.ts, content-sync/{sync,tarball,
validate}.ts, scripts/compile-content.ts + tests.

Gate: git grep content-sync/(github|publish-pin|drift|content-commit) -> zero;
tsc green; 113 relocated/affected tests pass.
