+++ apps/web/src/app/api/admin/content/drift/__tests__/route.test.ts
@@ -8,7 +8,7 @@ vi.mock("@/lib/admin/auth", () => ({
   adminUnauthorizedResponse: () => new Response("{}", { status: 401 }),
   requireAdminAuth: vi.fn(),
 }));
-vi.mock("@/lib/content-sync/github", () => ({
+vi.mock("@/lib/github/github", () => ({
   createGitHubClient: () => ({
     fetchHeadSha: async () => headSha,
     fetchChecksState: async () => "success",
@@ -76,7 +76,10 @@ describe("GET /api/admin/content/drift", () => {
     headSha = "c".repeat(40);
     courses = [{ _id: "course-x", title: "Course X" }];
     slots = new Map<string, SlotsLockT>([
-      ["course-x", { version: 1, slots: { "lesson-a": 0 }, retired: [], next: 1 }],
+      [
+        "course-x",
+        { version: 1, slots: { "lesson-a": 0 }, retired: [], next: 1 },
+      ],
     ]);
 
     const res = await get();
