+++ apps/web/src/app/api/admin/content/drift/route.ts
@@ -17,13 +17,10 @@ import {
 } from "@/lib/admin/sync-diff";
 import { slotsByCourseId } from "@/lib/content/store";
 import { SYNCED_SHA } from "@/lib/content/meta";
-import { createGitHubClient } from "@/lib/content-sync/github";
-import { deriveActiveMask } from "@/lib/content-sync/content-commit";
-import {
-  computeContentDrift,
-  computeChainDrift,
-} from "@/lib/content-sync/drift";
-import { GitHubUnavailableError } from "@/lib/content-sync/types";
+import { createGitHubClient } from "@/lib/github/github";
+import { deriveActiveMask } from "@/lib/github/content-commit";
+import { computeContentDrift, computeChainDrift } from "@/lib/github/drift";
+import { GitHubUnavailableError } from "@/lib/github/types";
 
 /**
  * The cheap chain-status a course needs for `computeChainDrift`. §11.1 replaces
